drive_api/requirements.txt
django>=3.1.2,<3.2.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers>=3.2.1,<3.3.0
pydrive>=1.3.1,<1.4.0